feat(graphql-rpc): add `Event.transactionBlock` field (#9616)
This patch exposes the parent transaction of an event through the
available event query paths.

As this introduces the possibility of recursive queries (e.g. `events ->
tranasctionBlock -> events -> ...`, new complexity limits are introduced
on the validation level exploiting
[https://async-graphql.github.io/async-graphql/en/depth_and_complexity.html](async-graphql
tooling) as opposed to the ad-hoc validation alredy present.

With the new limits, queries that try to recurse unnecessarily through
the available query paths are rejected as "Too complex".

Resolves #9402

Describe the tests that you ran to verify your changes.

Make sure to provide instructions for the maintainer as well as any
relevant configurations.

- [x] Basic tests (linting, compilation, formatting, unit/integration
tests)
- [x] Patch-specific tests (correctness, functionality coverage)

@iotaledger/infrastructure)

- [ ] Synchronization of the indexer from genesis for a network
including migration objects.
- [ ] Restart of indexer synchronization locally without resetting the
database.
- [ ] Restart of indexer synchronization on a production-like database.
- [ ] Deployment of services using Docker.
- [x] Verification of API backward compatibility.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [x] GraphQL: :bulb: A new field `Event.transactionBlock` is introduced
that exposes the parent transaction of an event.
- [ ] CLI:
- [ ] Rust SDK:
- [ ] REST API: